fix cypress tests bazel
