+++ integration/cypress.config.ts
@@ -1,17 +1,5 @@
 import {defineConfig} from 'cypress';
 
-declare global {
-  namespace Cypress {
-    interface Chainable<Subject> {
-      createChatPluginConversation(): Chainable<Subject>;
-      editContactDetails(): Chainable<Subject>;
-      connectChatPluginConnector(): Chainable<Subject>;
-      editDisplayName(): Chainable<Subject>;
-      listConversationsForContact(): Chainable<Subject>;
-    }
-  }
-}
-
 export default defineConfig({
   e2e: {
     baseUrl: 'http://localhost',
